+++ pkgs/swiftgen/test/integration/callbacks_bindings.dart
@@ -12,7 +12,7 @@ import 'dart:ffi' as ffi;
 import 'package:objective_c/objective_c.dart' as objc;
 import 'package:ffi/ffi.dart' as pkg_ffi;
 
-const _$objcVersionCheck = objc.ObjCVersionCheck(9, 5);
+const _$objcVersionCheck = objc.ObjCVersionCheck(9, 6);
 @ffi.Native<
   ffi.Pointer<objc.ObjCBlockImpl> Function(
     ffi.Int64,